requirements.txt
Django==4.1.3
pytest==7.2.0
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-mock==3.10.0
pylama==8.4.1
pytest-bdd==6.1.1
requests-mock==1.10.0
